backend/app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-multipart
opencv-python
opencv-python-headless
numpy
insightface
onnxruntime
pillow
